makefile

+++ Makefile
@@ -22,4 +22,12 @@ $(BIN): $(OBJ)
 	$(CC) $(CFLAGS) $(CPPFLAGS) $^ -o $@
 
 clean:
+ifeq ($(OS), Windows_NT)
+	powershell remove-item * -include y.tab.c, y.tab.h, lex.yy.c
+else
 	rm $(OBJ) y.tab.c y.tab.h lex.yy.c
+endif
+
+
+
+
